feat(cli): move download from repl to cli
